+++ tests/deprecated_api/test_remove_1-3.py
@@ -15,25 +15,10 @@
 
 import pytest
 
-from pytorch_lightning import LightningModule, Trainer
-from pytorch_lightning.callbacks import EarlyStopping, ModelCheckpoint
+from pytorch_lightning import LightningModule
 
 
 def test_v1_3_0_deprecated_arguments(tmpdir):
-    with pytest.deprecated_call(match='will no longer be supported in v1.3'):
-        callback = ModelCheckpoint()
-        Trainer(checkpoint_callback=callback, callbacks=[], default_root_dir=tmpdir)
-
-    # Deprecate prefix
-    with pytest.deprecated_call(match='will be removed in v1.3'):
-        ModelCheckpoint(prefix='temp')
-
-    # Deprecate auto mode
-    with pytest.deprecated_call(match='will be removed in v1.3'):
-        ModelCheckpoint(mode='auto')
-
-    with pytest.deprecated_call(match='will be removed in v1.3'):
-        EarlyStopping(mode='auto')
 
     with pytest.deprecated_call(match="The setter for self.hparams in LightningModule is deprecated"):